+++ plugins/main/server/lib/generate-alerts/types.d.ts
@@ -102,12 +102,10 @@ type DataKeys =
   | 'office365'
   | 'gcp'
   | 'audit'
-  | 'oscap'
   | 'title'
   | 'file'
   | 'virustotal'
   | 'vulnerability'
-  | 'osquery'
   | 'srcip'
   | 'srcuser'
   | 'srcport'
@@ -173,14 +171,10 @@ export type Params = {
   gcp: boolean;
   /* if true, set System Auditing fields */
   audit: boolean;
-  /* if true, set CIS-CAT fields */
-  ciscat: boolean;
   /* if true, set Docker fields */
   docker: boolean;
   /* if true, set Mitre att&ck fields */
   mitre: boolean;
-  /* if true, set OpenSCAP fields */
-  openscap: boolean;
   /* if true, set Policy monitoring fields */
   rootcheck: boolean;
   /* if true, set integrity monitoring fields (IMF) */
@@ -189,8 +183,6 @@ export type Params = {
   virustotal: boolean;
   /* if true, set vulnerabilities fields */
   vulnerabilities: boolean;
-  /* if true, set Osquery fields */
-  osquery: boolean;
   /* if true, set pci_dss fields */
   pci_dss: boolean;
   /* if true, set gdpr fields */